+++ cli/integration-test/integration/initialize_test.clj
@@ -79,7 +79,8 @@
                                                        "class"
                                                        "variable"
                                                        "method"
-                                                       "event"]
+                                                       "event"
+                                                       "interface"]
                                           :tokenModifiers ["definition"
                                                            "defaultLibrary"
                                                            "implementation"]}