+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Editor
   - Fix add missing import code action when there are multiple options. #1422
   - Only show completion items if no changed code is being processed, avoiding wrong completion items. #1425
+  - Improve semantic tokens for defprotocol, definterface coloring as `interface` tokens.
   
 - API/CLI
   - Make diagnostics command print as json or edn. #1419